# Remove `TOTAL_MEMORY_USAGE` and improve allocation performance by 1%