`TOTAL_MEMORY_USAGE` is a global variable, we use for tracking total memory usage.

Here are performance stats before the change:
```
alloc_32                time:   [38.936 ns 39.018 ns 39.143 ns]       
alloc_32                time:   [38.574 ns 38.633 ns 38.701 ns] 
alloc_32                time:   [38.511 ns 38.558 ns 38.606 ns]       
```
and after
```
alloc_32                
                        time:   [37.903 ns 37.935 ns 37.971 ns]                      
                        change: [-1.7741% -1.4475% -1.1053%] (p = 0.00 < 0.05)
                        Performance has improved.
alloc_32          time:   [38.101 ns 38.132 ns 38.162 ns]    
alloc_32           time:   [38.323 ns 38.359 ns 38.392 ns]              

```

There was an improvement of `0.5ns`, which is around `1-1.5%`.