@pmnoxx 2021-12-16
@matklad Criterion, does work. It's able to detect even small performance improvements.